finished lab

+++ run.sh
@@ -6,6 +6,11 @@ docker stop lab_container
 
 set -ex
 
+if [ ! -f "./run_$target.sql" ]; then
+	echo "Wrong target"
+	exit 1
+fi
+
 docker run --rm -d \
 -p 6432:5432 \
 --name lab_container \
@@ -16,13 +21,5 @@ postgres
 sleep 2
 
 psql -h localhost -p 6432 -U postgres -f ./init_db.sql
-if [[ "$target" = "procedures" ]]; then
-	psql -h localhost -p 6432 -U postgres -d lab_db -f ./run_procedures.sql
-elif [[ "$target" = "selects" ]]; then
-	psql -h localhost -p 6432 -U postgres -d lab_db -f ./run_selects.sql
-else
-	echo "Please, set target (procedures or selects)"
-	exit 1
-fi
-
+psql -h localhost -p 6432 -U postgres -d lab_db -f "./run_$target.sql"
 psql -h localhost -p 6432 -U postgres -d lab_db
